routes short and page num add
