+++ app.js
@@ -2,28 +2,24 @@ var createError = require('http-errors');
 var express = require('express');
 var path = require('path');
 var cookieParser = require('cookie-parser');
-var logger = require('morgan');
+var morgan = require('morgan');
 var session = require('express-session');
-
-
 
 var indexRouter = require('./routes/index');
 var loginRouter = require('./routes/login');
 var boardRouter = require('./routes/board');
 var introRouter = require('./routes/intro');
 var galleryRouter = require('./routes/gallery');
-var signUpRouter = require('./routes/signUp');
 var memberRouter = require('./routes/member');
-var findIdRouter = require('./routes/findId');
-var findPwRouter = require('./routes/findPw');
-var noticeRouter = require('./routes/notice');
+var researchRouter = require('./routes/research');
+var helmet = require('helmet')
 var app = express();
 
 // view engine setup
 app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'ejs');
 
-app.use(logger('dev'));
+app.use(morgan('dev'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
@@ -31,6 +27,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 //로그인 세션
 app.use(session({
+  name: 'sessionID',
   secret: 'test@~',
   resave: false,
   saveUninitialized: true,
@@ -44,13 +41,8 @@ app.use('/login', loginRouter);
 app.use('/board', boardRouter);
 app.use('/gallery', galleryRouter);
 app.use('/intro', introRouter);
-app.use('/signUp', signUpRouter);
 app.use('/member', memberRouter);
-app.use('/findId', findIdRouter);
-app.use('/findPw', findPwRouter);
-app.use('/notice', noticeRouter);
-
-
+app.use('/research', researchRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
@@ -67,6 +59,9 @@ app.use(function(err, req, res, next) {
   res.status(err.status || 500);
   res.render('error');
 });
+
+
+app.use(helmet());
 
 app.listen(3000);
 module.exports = app;
