+++ config/db.js
@@ -1,12 +1,13 @@
 const { TooManyRequests } = require('http-errors')
 const mysql = require('mysql')
+require('dotenv').config();
 
 const db_config = {
-    host: 'conative.myds.me',
-    port: '4304',
-    user : 'root',
-    password: 'qwerty12',
-    database: '5g_db'
+    host : process.env.DB_host,
+    port : process.env.DB_port,
+    user : process.env.DB_user,
+    password : process.env.DB_password,
+    database : process.env.DB_database
 };
 
 const db = mysql.createConnection(db_config);
