+++ package.json
@@ -9,11 +9,13 @@
     "cookie-parser": "~1.4.4",
     "dayjs": "^1.8.36",
     "debug": "~2.6.9",
+    "dotenv": "^8.2.0",
     "ejs": "~2.6.1",
     "express": "~4.16.1",
     "express-session": "^1.17.1",
+    "helmet": "^4.3.1",
     "http-errors": "~1.6.3",
-    "morgan": "~1.9.1",
+    "morgan": "^1.9.1",
     "mysql": "^2.18.1",
     "nodemon": "^2.0.6"
   }
