+++ routes/android.js
@@ -1,8 +1,10 @@
 var express = require('express');
 var router = express.Router();
+var db = require('../config/db')
+var dayjs =  require('dayjs')
 
 router.get('/', function(req, res, next) {
-    res.render('notice');
+    res.send('android connect');
 });
 
 module.exports = router;
