+++ routes/gallery.js
@@ -1,15 +1,16 @@
 var express = require('express');
 var router = express.Router();
+var bodyParser = require('body-parser')
 var db = require('../config/db')
 var dayjs =  require('dayjs')
 
-router.get('/', function(req, res, next) {
+router.get('/:num', function(req, res, next) {
     db.query('select * from Gallery', function (error, result) {
         if (error) {
             throw error;
           }    
         else {
-            res.render('gallery', {result : result, dayjs});
+            res.render('gallery', {result : result, g_num : req.params.num, max_value:9, dayjs });
             console.log(result);
         };
     });
@@ -21,7 +22,7 @@ router.get('/detail', function(req, res, next) {
             throw error;
           }    
         else {
-            res.render('gallery_detail', {result : result, dayjs});
+            res.render('gallery', {result : result, g_num : req.params.num, max_value:9, dayjs });
             console.log(result);
         };
     });
