+++ routes/intro.js
@@ -4,5 +4,14 @@ var router = express.Router();
 router.get('/', function(req, res, next) {
     res.render('intro');
 });
+router.get('/greetings', function(req, res, next) {
+    res.render('intro');
+});
+router.get('/history', function(req, res, next) {
+    res.render('intro');
+});
+router.get('/organization', function(req, res, next) {
+    res.render('intro');
+});
 
 module.exports = router;
