+++ views/board.ejs
@@ -25,19 +25,19 @@
     </ol>
     <div class="carousel-inner">
       <div class="carousel-item active">
-        <img src="images/5g.jpg" class="d-block w-100" alt="...">
+        <img src="/images/5g.jpg" class="d-block w-100" alt="...">
         <div class="carousel-caption d-none d-md-block">
           <h5>First slide label</h5>
         </div>
       </div>
       <div class="carousel-item">
-        <img src="images/5g.jpg" class="d-block w-100" alt="...">
+        <img src="/images/5g.jpg" class="d-block w-100" alt="...">
         <div class="carousel-caption d-none d-md-block">
           <h5>Second slide label</h5>
         </div>
       </div>
       <div class="carousel-item">
-        <img src="images/5g.jpg" class="d-block w-100" alt="...">
+        <img src="/images/5g.jpg" class="d-block w-100" alt="...">
         <div class="carousel-caption d-none d-md-block">
           <h5>Third slide label</h5>
         </div>
