+++ views/findId.ejs
@@ -17,12 +17,12 @@
             integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV"
             crossorigin="anonymous"></script>
     
-        <link href="./stylesheets/style.css" rel="stylesheet">
+        <link href="/stylesheets/style.css" rel="stylesheet">
     
     </head>
     <body class="text-center bg">
         <form action="./findId" method="POST" class="form-signin">
-            <img class="mb-4" src="./images/5g.png" alt="" width="72" height="72">
+            <img class="mb-4" src="/images/5g.png" alt="" width="72" height="72">
             <h3 style="font-family: maple;">ID찾기</h1>
                 <input type="text" name="input_email" class="form-control" placeholder="email">
