+++ views/find_Change_Pw.ejs
@@ -11,7 +11,7 @@
 <body>
    
     당신의 새로운 비밀번호를 입력해주세요. <br>
-    <form action="/findPw/pwCheck_detail" method="POST" class="form-pwcheck">
+    <form action="./pwCheck" method="POST" class="form-pwcheck">
     <input type="text" name="input_email" class="form-control mt-3 mb-1" placeholder="EMAIL"><br>
     <input type="password" name="changepw" class="form-control mt-3 mb-1" placeholder="PW"><br>
     <input type="password" name="changepw_check" class="form-control" placeholder="PWCHECK"><br>
