+++ views/gallery.ejs
@@ -19,7 +19,7 @@
     <title>5G navbar</title>
 
     <style>
-       @import url(./stylesheets/album.css);
+       @import url(/stylesheets/album.css);
     </style>
 </head>
 
@@ -32,7 +32,7 @@
         <div style="margin-top:60px; margin-bottom: 30px" id="carouselExampleFade" class="carousel slide carousel-fade" data-ride="carousel">
             <div class="carousel-inner">
                 <div class="carousel-item active" data-interval="2000">
-                    <img src="./images/gall.jpg" class="d-block w-100" alt="0" style="height: 40vh;">
+                    <img src="/images/gall.jpg" class="d-block w-100" alt="0" style="height: 40vh;">
                 </div>
             </div>
         </div>
@@ -40,29 +40,43 @@
         <div class="album py-5">
             <div class="container">
                 <div class="row">
-                    <% for(let i= 0; i< result.length; i++){ %>
-                    <div class="col-md-4" style="margin-bottom: 30px;">
-                        <div class="card" style="border: solid 1px black;">
-                            <a href="#" onclick="show()">
-                                <img src="./images/<%=result[i].g_img%>" alt="" width="100%" height="300">
-                                <div class="card-body">
-                                    <p class="card-text" style="color: black;"><%=result[i].g_title%></p>
-                                    <hr>
-                                    <p class="card-text" style="color: black; text-align: right;"><%= dayjs(result[i].g_write_date).format('YYYY.MM'); %></p>
-                                </div>
-                            </a>
+                    <% for(var i = (g_num - 1) * max_value; i < g_num * max_value; i++){ %>
+                        <% var data_value = result[i];
+                            if(data_value === undefined){
+                                continue;
+                            }
+                        %>
+                        <div class="col-md-4" style="margin-bottom: 30px;">
+                            <div class="card" style="border: solid 1px black;">
+                                <a href="#" onclick="show()">
+                                    <img src="/images/<%= data_value.g_img%>" alt="" width="100%" height="300">
+                                    <div class="card-body">
+                                        <p class="card-text" style="color: black;"><%=data_value.g_title%></p><hr>
+                                        <p class="card-text" style="color: black; text-align: right;"><%= dayjs(data_value.g_write_date).format('YYYY.MM'); %></p>
+                                    </div></a>
+                            </div>
                         </div>
-                    </div>
                     <% } %>
                 </div>
+                <div>
+                    <nav>
+                      <ul class="pagination justify-content-center">
+                        <li class="page-item"><a class="page-link" style="color: black;" href="#">이전</a></li>
+                        <% for(var i = 1; i<= Math.ceil(result.length / max_value); i++){ %>
+                        <li class="page-item"><a class="page-link" style="color: black;" href="/gallery/<%= i %>"><%= i %></a>
+                        <% } %>
+                        <li class="page-item"><a class="page-link"style="color: black;" href="#">다음</a></li>
+                      </ul>
+                    </nav>
+                  </div>
             </div>
         </div>
     </main>
 
     <div class="pop" id="modal">
         <div class="view">
-            <a href="#" onclick="hide()"><img src="./images/exit.png" alt="" class="prev"></a>
-            <img src="./images/gall.jpg" alt="" class="picture">
+            <a href="#" onclick="hide()"><img src="/images/exit.png" alt="" class="prev"></a>
+            <img src="/images/gall.jpg" alt="" class="picture">
             <div class="text_box">
                 제목<br>
                 내용<br>
