+++ views/index.ejs
@@ -19,14 +19,14 @@
     <style>
         @font-face {
             font-family: maple;
-            src: url('./font/Maplestory Bold.ttf');
+            src: url('/font/Maplestory Bold.ttf');
         }
 
         @import url(./stylesheets/index.css);
     </style>
 
     <title>5G</title>
-    <link href="./images/5g.png" rel="shortcut icon" type="image/x-icon">
+    <link href="/images/5g.png" rel="shortcut icon" type="image/x-icon">
 </head>
 
 <body>
@@ -41,7 +41,7 @@
             </ol>
             <div class="carousel-inner">
                 <div class="carousel-item active" data-interval="2000">
-                    <img src="./images/555.jpg" class="d-block w-100" alt="0" style="height: 50vh;">
+                    <img src="/images/555.jpg" class="d-block w-100" alt="0" style="height: 50vh;">
                 </div>
             </div>
             <a class="carousel-control-prev" href="#carouselExampleFade" role="button" data-slide="prev">
@@ -72,7 +72,7 @@
                         <tr>
                             <td>
                                 <span>
-                                    <a href="#"><img src="./images/itrc.png" style="width: 100px; height: 100px; margin-right: 15px"></a>
+                                    <a href="#"><img src="/images/itrc.png" style="width: 100px; height: 100px; margin-right: 15px"></a>
                                 </span>
                             </td>
                             <td>
@@ -89,7 +89,7 @@
                         <tr>
                             <td>
                                 <span>
-                                    <a href="#"><img src="./images/logo.png" style="width: 100px; height: 100px; margin-right: 15px"></a>
+                                    <a href="#"><img src="/images/logo.png" style="width: 100px; height: 100px; margin-right: 15px"></a>
                                 </span>
                             </td>
                             <td>
@@ -106,7 +106,7 @@
                         <tr>
                             <td>
                                 <span>
-                                    <a href="#"><img src="./images/logo.png" style="width: 100px; height: 100px; margin-right: 15px"></a>
+                                    <a href="#"><img src="/images/logo.png" style="width: 100px; height: 100px; margin-right: 15px"></a>
                                 </span>
                             </td>
                             <td>
@@ -134,7 +134,7 @@
                         <tr>
                             <td>
                                 <span>
-                                    <a href="#"><img src="./images/report.svg" style="width: 100px; height: 100px; margin-right: 15px"></a>
+                                    <a href="#"><img src="/images/report.svg" style="width: 100px; height: 100px; margin-right: 15px"></a>
                                 </span>
                             </td>
                             <td>
@@ -151,7 +151,7 @@
                         <tr>
                             <td>
                                 <span>
-                                    <a href="#"><img src="./images/report.svg" style="width: 100px; height: 100px; margin-right: 15px"></a>
+                                    <a href="#"><img src="/images/report.svg" style="width: 100px; height: 100px; margin-right: 15px"></a>
                                 </span>
                             </td>
                             <td>
@@ -168,7 +168,7 @@
                         <tr>
                             <td>
                                 <span>
-                                    <a href="#"><img src="./images/report.svg" style="width: 100px; height: 100px; margin-right: 15px"></a>
+                                    <a href="#"><img src="/images/report.svg" style="width: 100px; height: 100px; margin-right: 15px"></a>
                                 </span>
                             </td>
                             <td>
@@ -197,7 +197,7 @@
                         <tr>
                             <td>
                                 <span>
-                                    <a href="#"><img src="./images/<%=result[i].g_img%>" style="width: 100px; height: 100px; margin-right: 15px"></a>
+                                    <a href="#"><img src="/images/<%=result[i].g_img%>" style="width: 100px; height: 100px; margin-right: 15px"></a>
                                 </span>
                             </td>
                             <td>
@@ -222,7 +222,7 @@
                 </p>
             </div>
             <div class="col-md-5">
-                <img src="./images/cat.jpg" class="img-fluid" width="500" height="500">
+                <img src="/images/cat.jpg" class="img-fluid" width="500" height="500">
             </div>
         </div>
