+++ views/intro.ejs
@@ -44,7 +44,7 @@
         <% include ./nav %>
     </header>
     <div class="jb-wrap">
-        <div class="jb-image"><img src="./images/intro.jpg" class="img-fluid" alt="Responsive image" style="width: 100%;"></div>
+        <div class="jb-image"><img src="/images/intro.jpg" class="img-fluid" alt="Responsive image" style="width: 100%;"></div>
         <div class="jb-text">해양IT 융합기술연구소는<br><br>
 
             전 세계적으로 매년 약 8,200만명의 인구가 증가하면서 향후 지구 전체 인구에 대한 식량, 에너지 및 자원 부족이 예상됩니다.<br><br>
