+++ views/nav.ejs
@@ -17,7 +17,7 @@
         crossorigin="anonymous"></script>
 
     <style>
-        @import url(./stylesheets/nav.css);
+        @import url(/stylesheets/nav.css);
     </style>
 </head>
 <body style="min-width: 770px;">
@@ -28,19 +28,19 @@
                     <a class="nav-link" href="/intro">대표말</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/member">멤버소개</a>
+                    <a class="nav-link" href="/member/1">멤버소개</a>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" href="/board">학술논문</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/gallery">갤러리</a>
+                    <a class="nav-link" href="/gallery/1">갤러리</a>
                 </li>
             </ul>
         </div>
         <span align=center class="mx-auto">
             <a class="navbar-brand" href="/">
-                <img src="./images/5g.png" class="logo_size" alt="">
+                <img src="/images/5g.png" class="logo_size" alt="">
             </a>
         </span>
         <div align=right class="collapse navbar-collapse w-100 order-3" id="navbar-main">
