+++ views/signUp.ejs
@@ -7,11 +7,11 @@
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js" integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6" crossorigin="anonymous"></script>
 
-    <link rel="stylesheet" href="./stylesheets/signup.css">
+    <link rel="stylesheet" href="../stylesheets/signup.css">
 </head>
 <body>
         <div id="logo" align="center">
-            <img src="images/5g.png" alt="">
+            <img src="/images/5g.png" alt="">
         </div>
 
         <div class="formsign">
